Aligned and Updated NavBar
Added MLH Banner on left side and changed nav content to aligned right side with tailored tags
