+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Footer from './component//Footer/Footer.jsx';
 import Main from './component/Main/Main.jsx'
-import NavigationBar from './component/Header/Navigation-Bar.jsx'
+import NavigationBar from './component/Header/NavigationBar.jsx'
 const App = () => {
   return (
     <React.Fragment> 		{/* This allows children, and removal of unnecessary <div> wrapper */}
