--- client/src/component/Header/Day-Banner.css
@@ -1,11 +1,0 @@
-.App-banner{
-    /*Adjusting position of the banner*/
-    display:block;
-    max-width:100px;
-    min-width:60px;
-    position:absolute;
-    right:84px;
-    top: 82px;
-    width:10%;
-    z-index:10000
-}
