+++ client/src/component/Header/navBar.css
@@ -1,32 +1,21 @@
 .App-nav {
-    background-color: #2a2834;
+  background-color: #2a2834; /* This adjusts the container */
+  min-height: 2vh;
+  padding-top: 1vh;
+  padding-bottom: 2vh;
+  display: flex;
+  justify-content: flex-end; /* makes them align to the right side */
 }
 
-.HackMerced-logo{
-   /* Adjusting the HackMerced logo */
-  position: relative;
-  left: 10px; /* positioning to correct location */
-  top: 4px;
-    max-height: 100%; /* adaptive height */
-      max-width: 100%; /* adaptive width */
-}
-
-.nav-link{ /* This adjusts the containers themselves */
-   display: inline-flex; /* makes containers inline horizontally */
-   width: 10.5%; /* 8 sections/100 adjusted for padding */
-   align-items: center; 
-   justify-content: center; /* centers items inside the divs */
-   font-size: 25px;
-   padding-left:12.5px;
-   padding-right:12.5px;
-   min-height: 8vh;
-   padding-bottom: 10px;
-   padding-top: 5px;
+.nav-link{ /* This adjusts for the content inside the container */
+  text-align: center;
+   width: 12%; 
+   font-size: 2vw;
 }
 
 /* I had to seperate the CSS for links and nav bars as the changes wouldn't work through the containers*/
 
-.style-link{ /* This adjusts the React Router links themselves */
+.style-link{ /* This adjusts the React Router links */
   color: white;
   text-decoration: none; /* removes underscores of the links */
 }
