+++ client/src/pages/home/home.jsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import logo from '../../logo.svg';
 import './home.css'; //import css for this page
-import DayBanner from '../../component/Header/Day-Banner.jsx';
 
 class Home extends React.Component {   // This page is the exact same page as the original example
   render() {
@@ -20,9 +19,8 @@ class Home extends React.Component {   // This page is the exact same page as th
             
             Learn React
             <React.Fragment> 
-              <DayBanner /> 
+
             </React.Fragment>
-          
           </a>
       </article>
       );
